Extract shared worker request pipeline from dev-entry and dynamic bootstrap
